fix : 리뷰 변경사항 수정
- long 대신 Wrapper타입 Long사용
- repository int -> Long으로 변

+++ haul-be/src/main/java/com/hansalchai/haul/driver/entity/Driver.java
@@ -9,5 +9,5 @@ import jakarta.persistence.Id;
 public class Driver {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int driverId;
+	private Long driverId;
 }
