+++ haul-be/src/main/java/com/hansalchai/haul/owner/repository/OwnerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hansalchai.haul.owner.entity.Owner;
 
-public interface OwnerRepository extends JpaRepository<Owner, Integer> {
+public interface OwnerRepository extends JpaRepository<Owner, Long> {
 
 }
