+++ haul-be/src/main/java/com/hansalchai/haul/reservation/repository/CargoOptionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hansalchai.haul.reservation.entity.CargoOption;
 
-public interface CargoOptionRepository extends JpaRepository<CargoOption, Integer> {
+public interface CargoOptionRepository extends JpaRepository<CargoOption, Long> {
 }
