+++ haul-be/src/main/java/com/hansalchai/haul/reservation/repository/CargoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hansalchai.haul.reservation.entity.Cargo;
 
-public interface CargoRepository extends JpaRepository<Cargo, Integer> {
+public interface CargoRepository extends JpaRepository<Cargo, Long> {
 }
