+++ haul-be/src/main/java/com/hansalchai/haul/reservation/repository/SourceRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hansalchai.haul.reservation.entity.Source;
 
-public interface SourceRepository extends JpaRepository<Source, Integer> {
+public interface SourceRepository extends JpaRepository<Source, Long> {
 }
